requirements.txt
appdirs==1.4.4
attrs==21.4.0
beautifulsoup4==4.11.1
cattrs==1.10.0
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
lxml==4.8.0
markdownify==0.11.2
MarkdownPP==1.5.1
pandoc==2.2
plumbum==1.7.2
ply==3.11
PyYAML==6.0
requests==2.27.1
requests-cache==0.9.4
six==1.16.0
soupsieve==2.3.2.post1
url-normalize==1.4.3
urllib3==1.26.9
watchdog==2.1.7
